cppad_py: switch to using pytest organization for tests.
functions.py: preliminary ad version of curvefit functions file.
